This commit was generated by cvs2svn to compensate for changes in r17, which included commits to RCS files with non-trunk default branches.

+++ mvdsv/source/gl_ngraph.c
@@ -105,7 +105,7 @@ void R_NetGraph (void)
 		for (x = 0; x < NET_TIMINGS; x++)
 			ngraph_pixels[y][x] = d_8to24table[ngraph_texels[y][x]];
 
-	x =	-((vid.width - 320)>>1);
+	x =	-(int)((vid.width - 320)>>1);
 	y = vid.height - sb_lines - 24 - NET_GRAPHHEIGHT - 1;
 
 	M_DrawTextBox (x, y, NET_TIMINGS/8, NET_GRAPHHEIGHT/8 + 1);
